requirements.txt
langchain
olefile
pywin32
pypdf
python-docx
python-pptx
chromadb
zipfile
openai
tiktoken